feat: createLogger for easier logging in individual modules
